+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-06-04T06:34:11.043Z")
+    created: Optional[datetime] = parse_datetime("2025-06-04T07:24:13.399Z")
     r"""The date and time the resource was created"""