+++ packages/orq-rc/RELEASES.md
@@ -1679,3 +1679,13 @@ Based on:
 - [python v3.8.0-rc.0] packages/orq-rc
 ### Releases
 - [PyPI v3.8.0-rc.0] https://pypi.org/project/orq-ai-sdk/3.8.0-rc.0 - packages/orq-rc
+
+## 2025-06-04 07:25:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.555.2 (2.620.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.8.0-rc.1] packages/orq-rc
+### Releases
+- [PyPI v3.8.0-rc.1] https://pypi.org/project/orq-ai-sdk/3.8.0-rc.1 - packages/orq-rc